Low-level API: Easier way to handle simplex/duplex endpoints
