+++ .pre-commit-config.yaml
@@ -13,49 +13,6 @@ repos:
     hooks:
       - id: pyupgrade
         args: ['--py311-plus']
-  - repo: local
-    hooks:
-      - id: mypy
-        name: mypy (project)
-        entry: tox run -q -e mypy-full
-        language: system
-        files: ^(src/)
-        types_or: [python, pyi]
-        require_serial: true
-        pass_filenames: false
-      - id: mypy
-        name: mypy (tests)
-        files: ^((src|tests)/)
-        entry: tox run -q -e mypy-test
-        language: system
-        types_or: [python, pyi]
-        require_serial: true
-        pass_filenames: false
-      - id: mypy
-        name: mypy (docs)
-        files: ^((src|docs/source)/)
-        exclude: ^(docs/source/conf.py)$
-        entry: tox run -q -e mypy-docs
-        language: system
-        types_or: [python, pyi]
-        require_serial: true
-        pass_filenames: false
-      - id: mypy
-        name: mypy (benchmark/servers)
-        files: ^((src|benchmark_server)/)
-        entry: tox run -q -e mypy-benchmark_server
-        language: system
-        types_or: [python, pyi]
-        require_serial: true
-        pass_filenames: false
-      - id: mypy
-        name: mypy (benchmark/micro-benchs)
-        files: ^((src|micro_benchmarks)/)
-        entry: tox run -q -e mypy-micro_benchmarks
-        language: system
-        types_or: [python, pyi]
-        require_serial: true
-        pass_filenames: false
   - repo: https://github.com/PyCQA/isort
     rev: '5.13.2'
     hooks:
@@ -111,3 +68,46 @@ repos:
       - id: check-json
       - id: check-toml
       - id: check-yaml
+  - repo: local
+    hooks:
+      - id: mypy
+        name: mypy (project)
+        entry: tox run -q -e mypy-full
+        language: system
+        files: ^(src/)
+        types_or: [python, pyi]
+        require_serial: true
+        pass_filenames: false
+      - id: mypy
+        name: mypy (tests)
+        files: ^((src|tests)/)
+        entry: tox run -q -e mypy-test
+        language: system
+        types_or: [python, pyi]
+        require_serial: true
+        pass_filenames: false
+      - id: mypy
+        name: mypy (docs)
+        files: ^((src|docs/source)/)
+        exclude: ^(docs/source/conf.py)$
+        entry: tox run -q -e mypy-docs
+        language: system
+        types_or: [python, pyi]
+        require_serial: true
+        pass_filenames: false
+      - id: mypy
+        name: mypy (benchmark/servers)
+        files: ^((src|benchmark_server)/)
+        entry: tox run -q -e mypy-benchmark_server
+        language: system
+        types_or: [python, pyi]
+        require_serial: true
+        pass_filenames: false
+      - id: mypy
+        name: mypy (benchmark/micro-benchs)
+        files: ^((src|micro_benchmarks)/)
+        entry: tox run -q -e mypy-micro_benchmarks
+        language: system
+        types_or: [python, pyi]
+        require_serial: true
+        pass_filenames: false
