+++ src/easynetwork/lowlevel/api_async/endpoints/datagram.py
@@ -16,40 +16,201 @@
 
 from __future__ import annotations
 
-__all__ = ["AsyncDatagramEndpoint"]
+__all__ = [
+    "AsyncDatagramEndpoint",
+    "AsyncDatagramReceiverEndpoint",
+    "AsyncDatagramSenderEndpoint",
+]
 
+import dataclasses
 import warnings
 from collections.abc import Callable, Mapping
-from typing import Any, Generic, TypeGuard
+from typing import Any, Generic
 
 from .... import protocol as protocol_module
 from ...._typevars import _T_ReceivedPacket, _T_SentPacket
-from ....exceptions import DatagramProtocolParseError, UnsupportedOperation
-from ... import _utils, typed_attr
+from ....exceptions import DatagramProtocolParseError
+from ... import _utils
 from ..backend.abc import AsyncBackend
 from ..transports import abc as transports
 
 
-class AsyncDatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPacket, _T_ReceivedPacket]):
+class AsyncDatagramReceiverEndpoint(transports.AsyncBaseTransport, Generic[_T_ReceivedPacket]):
     """
-    A communication endpoint based on unreliable packets of data.
+    A read-only communication endpoint based on unreliable packets of data.
     """
 
     __slots__ = (
         "__transport",
-        "__is_read_transport",
-        "__is_write_transport",
-        "__protocol",
-        "__send_guard",
+        "__receiver",
         "__recv_guard",
-        "__weakref__",
     )
 
     def __init__(
         self,
-        transport: (
-            transports.AsyncDatagramTransport | transports.AsyncDatagramReadTransport | transports.AsyncDatagramWriteTransport
-        ),
+        transport: transports.AsyncDatagramReadTransport,
+        protocol: protocol_module.DatagramProtocol[Any, _T_ReceivedPacket],
+    ) -> None:
+        """
+        Parameters:
+            transport: The data transport to use.
+            protocol: The :term:`protocol object` to use.
+        """
+
+        if not isinstance(transport, transports.AsyncDatagramReadTransport):
+            raise TypeError(f"Expected an AsyncDatagramReadTransport object, got {transport!r}")
+        if not isinstance(protocol, protocol_module.DatagramProtocol):
+            raise TypeError(f"Expected a DatagramProtocol object, got {protocol!r}")
+
+        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] = _DataReceiverImpl(transport, protocol)
+
+        self.__transport: transports.AsyncDatagramReadTransport = transport
+        self.__recv_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently receving data on this endpoint")
+
+    def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
+        try:
+            transport = self.__transport
+        except AttributeError:
+            return
+        if not transport.is_closing():
+            msg = f"unclosed endpoint {self!r} pointing to {transport!r} (and cannot be closed synchronously)"
+            _warn(msg, ResourceWarning, source=self)
+
+    def is_closing(self) -> bool:
+        """
+        Checks if the endpoint is closed or in the process of being closed.
+
+        Returns:
+            :data:`True` if the endpoint is closed.
+        """
+        return self.__transport.is_closing()
+
+    async def aclose(self) -> None:
+        """
+        Closes the endpoint.
+        """
+        await self.__transport.aclose()
+
+    async def recv_packet(self) -> _T_ReceivedPacket:
+        """
+        Waits for a new packet from the remote endpoint.
+
+        Raises:
+            DatagramProtocolParseError: invalid data received.
+
+        Returns:
+            the received packet.
+        """
+        with self.__recv_guard:
+            receiver = self.__receiver
+
+            return await receiver.receive()
+
+    @_utils.inherit_doc(transports.AsyncBaseTransport)
+    def backend(self) -> AsyncBackend:
+        return self.__transport.backend()
+
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
+
+
+class AsyncDatagramSenderEndpoint(transports.AsyncBaseTransport, Generic[_T_SentPacket]):
+    """
+    A write-only communication endpoint based on unreliable packets of data.
+    """
+
+    __slots__ = (
+        "__transport",
+        "__sender",
+        "__send_guard",
+    )
+
+    def __init__(
+        self,
+        transport: transports.AsyncDatagramWriteTransport,
+        protocol: protocol_module.DatagramProtocol[_T_SentPacket, Any],
+    ) -> None:
+        """
+        Parameters:
+            transport: The data transport to use.
+            protocol: The :term:`protocol object` to use.
+        """
+
+        if not isinstance(transport, transports.AsyncDatagramWriteTransport):
+            raise TypeError(f"Expected an AsyncDatagramWriteTransport object, got {transport!r}")
+        if not isinstance(protocol, protocol_module.DatagramProtocol):
+            raise TypeError(f"Expected a DatagramProtocol object, got {protocol!r}")
+
+        self.__sender: _DataSenderImpl[_T_SentPacket] = _DataSenderImpl(transport, protocol)
+
+        self.__transport: transports.AsyncDatagramWriteTransport = transport
+        self.__send_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently sending data on this endpoint")
+
+    def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
+        try:
+            transport = self.__transport
+        except AttributeError:
+            return
+        if not transport.is_closing():
+            msg = f"unclosed endpoint {self!r} pointing to {transport!r} (and cannot be closed synchronously)"
+            _warn(msg, ResourceWarning, source=self)
+
+    def is_closing(self) -> bool:
+        """
+        Checks if the endpoint is closed or in the process of being closed.
+
+        Returns:
+            :data:`True` if the endpoint is closed.
+        """
+        return self.__transport.is_closing()
+
+    async def aclose(self) -> None:
+        """
+        Closes the endpoint.
+        """
+        await self.__transport.aclose()
+
+    async def send_packet(self, packet: _T_SentPacket) -> None:
+        """
+        Sends `packet` to the remote endpoint.
+
+        Warning:
+            In the case of a cancellation, it is impossible to know if all the packet data has been sent.
+
+        Parameters:
+            packet: the Python object to send.
+        """
+        with self.__send_guard:
+            sender = self.__sender
+
+            await sender.send(packet)
+
+    @_utils.inherit_doc(transports.AsyncBaseTransport)
+    def backend(self) -> AsyncBackend:
+        return self.__transport.backend()
+
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
+
+
+class AsyncDatagramEndpoint(transports.AsyncBaseTransport, Generic[_T_SentPacket, _T_ReceivedPacket]):
+    """
+    A full-duplex communication endpoint based on unreliable packets of data.
+    """
+
+    __slots__ = (
+        "__transport",
+        "__sender",
+        "__receiver",
+        "__send_guard",
+        "__recv_guard",
+    )
+
+    def __init__(
+        self,
+        transport: transports.AsyncDatagramTransport,
         protocol: protocol_module.DatagramProtocol[_T_SentPacket, _T_ReceivedPacket],
     ) -> None:
         """
@@ -58,15 +219,15 @@ class AsyncDatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPa
             protocol: The :term:`protocol object` to use.
         """
 
-        if not isinstance(transport, (transports.AsyncDatagramReadTransport, transports.AsyncDatagramWriteTransport)):
+        if not isinstance(transport, transports.AsyncDatagramTransport):
             raise TypeError(f"Expected an AsyncDatagramTransport object, got {transport!r}")
         if not isinstance(protocol, protocol_module.DatagramProtocol):
             raise TypeError(f"Expected a DatagramProtocol object, got {protocol!r}")
 
-        self.__is_read_transport: bool = isinstance(transport, transports.AsyncDatagramReadTransport)
-        self.__is_write_transport: bool = isinstance(transport, transports.AsyncDatagramWriteTransport)
-        self.__transport: transports.AsyncDatagramReadTransport | transports.AsyncDatagramWriteTransport = transport
-        self.__protocol: protocol_module.DatagramProtocol[_T_SentPacket, _T_ReceivedPacket] = protocol
+        self.__sender: _DataSenderImpl[_T_SentPacket] = _DataSenderImpl(transport, protocol)
+        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] = _DataReceiverImpl(transport, protocol)
+
+        self.__transport: transports.AsyncDatagramTransport = transport
         self.__send_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently sending data on this endpoint")
         self.__recv_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently receving data on this endpoint")
 
@@ -105,20 +266,9 @@ class AsyncDatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPa
             packet: the Python object to send.
         """
         with self.__send_guard:
-            transport = self.__transport
-            protocol = self.__protocol
+            sender = self.__sender
 
-            if not self.__supports_write(transport):
-                raise UnsupportedOperation("transport does not support sending data")
-
-            try:
-                datagram: bytes = protocol.make_datagram(packet)
-            except Exception as exc:
-                raise RuntimeError("protocol.make_datagram() crashed") from exc
-            finally:
-                del packet
-
-            await transport.send(datagram)
+            await sender.send(packet)
 
     async def recv_packet(self) -> _T_ReceivedPacket:
         """
@@ -131,32 +281,47 @@ class AsyncDatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPa
             the received packet.
         """
         with self.__recv_guard:
-            transport = self.__transport
-            protocol = self.__protocol
+            receiver = self.__receiver
 
-            if not self.__supports_read(transport):
-                raise UnsupportedOperation("transport does not support receiving data")
-
-            datagram = await transport.recv()
-            try:
-                return protocol.build_packet_from_datagram(datagram)
-            except DatagramProtocolParseError:
-                raise
-            except Exception as exc:
-                raise RuntimeError("protocol.build_packet_from_datagram() crashed") from exc
-            finally:
-                del datagram
+            return await receiver.receive()
 
     @_utils.inherit_doc(transports.AsyncBaseTransport)
     def backend(self) -> AsyncBackend:
         return self.__transport.backend()
 
-    def __supports_read(self, transport: transports.AsyncBaseTransport) -> TypeGuard[transports.AsyncDatagramReadTransport]:
-        return self.__is_read_transport
-
-    def __supports_write(self, transport: transports.AsyncBaseTransport) -> TypeGuard[transports.AsyncDatagramWriteTransport]:
-        return self.__is_write_transport
-
     @property
     def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
         return self.__transport.extra_attributes
+
+
+@dataclasses.dataclass(slots=True)
+class _DataSenderImpl(Generic[_T_SentPacket]):
+    transport: transports.AsyncDatagramWriteTransport
+    protocol: protocol_module.DatagramProtocol[_T_SentPacket, Any]
+
+    async def send(self, packet: _T_SentPacket) -> None:
+        try:
+            datagram: bytes = self.protocol.make_datagram(packet)
+        except Exception as exc:
+            raise RuntimeError("protocol.make_datagram() crashed") from exc
+        finally:
+            del packet
+
+        await self.transport.send(datagram)
+
+
+@dataclasses.dataclass(slots=True)
+class _DataReceiverImpl(Generic[_T_ReceivedPacket]):
+    transport: transports.AsyncDatagramReadTransport
+    protocol: protocol_module.DatagramProtocol[Any, _T_ReceivedPacket]
+
+    async def receive(self) -> _T_ReceivedPacket:
+        datagram = await self.transport.recv()
+        try:
+            return self.protocol.build_packet_from_datagram(datagram)
+        except DatagramProtocolParseError:
+            raise
+        except Exception as exc:
+            raise RuntimeError("protocol.build_packet_from_datagram() crashed") from exc
+        finally:
+            del datagram
