+++ src/easynetwork/lowlevel/api_async/endpoints/stream.py
@@ -16,7 +16,11 @@
 
 from __future__ import annotations
 
-__all__ = ["AsyncStreamEndpoint"]
+__all__ = [
+    "AsyncStreamEndpoint",
+    "AsyncStreamReceiverEndpoint",
+    "AsyncStreamSenderEndpoint",
+]
 
 import dataclasses
 import errno as _errno
@@ -28,14 +32,204 @@ from .... import protocol as protocol_module
 from ...._typevars import _T_ReceivedPacket, _T_SentPacket
 from ....exceptions import UnsupportedOperation
 from ....warnings import ManualBufferAllocationWarning
-from ... import _stream, _utils, typed_attr
+from ... import _stream, _utils
 from ..backend.abc import AsyncBackend
 from ..transports import abc as transports
 
 
-class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPacket, _T_ReceivedPacket]):
+class AsyncStreamReceiverEndpoint(transports.AsyncBaseTransport, Generic[_T_ReceivedPacket]):
     """
-    A communication endpoint based on continuous stream data transport.
+    A read-only communication endpoint based on continuous stream data transport.
+    """
+
+    __slots__ = (
+        "__transport",
+        "__receiver",
+        "__recv_guard",
+    )
+
+    def __init__(
+        self,
+        transport: transports.AsyncStreamReadTransport,
+        protocol: protocol_module.StreamProtocol[Any, _T_ReceivedPacket],
+        max_recv_size: int,
+        *,
+        manual_buffer_allocation: Literal["try", "no", "force"] = "try",
+        manual_buffer_allocation_warning_stacklevel: int = 2,
+    ) -> None:
+        """
+        Parameters:
+            transport: The data transport to use.
+            protocol: The :term:`protocol object` to use.
+            max_recv_size: Read buffer size.
+            manual_buffer_allocation: Select whether or not to enable the manual buffer allocation system:
+
+                                      * ``"try"``: (the default) will use the buffer API if the transport and protocol support it,
+                                        and fall back to the default implementation otherwise.
+                                        Emits a :exc:`.ManualBufferAllocationWarning` if only the transport does not support it.
+
+                                      * ``"no"``: does not use the buffer API, even if they both support it.
+
+                                      * ``"force"``: requires the buffer API. Raises :exc:`.UnsupportedOperation` if it fails and
+                                        no warnings are emitted.
+            manual_buffer_allocation_warning_stacklevel: ``stacklevel`` parameter of :func:`warnings.warn` when emitting
+                                                         :exc:`.ManualBufferAllocationWarning`.
+        """
+
+        if not isinstance(transport, transports.AsyncStreamReadTransport):
+            raise TypeError(f"Expected an AsyncStreamReadTransport object, got {transport!r}")
+        _check_max_recv_size_value(max_recv_size)
+        _check_manual_buffer_allocation_value(manual_buffer_allocation)
+        manual_buffer_allocation_warning_stacklevel = max(manual_buffer_allocation_warning_stacklevel, 2)
+
+        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] | _BufferedReceiverImpl[_T_ReceivedPacket] = _get_receiver(
+            transport=transport,
+            protocol=protocol,
+            max_recv_size=max_recv_size,
+            manual_buffer_allocation=manual_buffer_allocation,
+            manual_buffer_allocation_warning_stacklevel=manual_buffer_allocation_warning_stacklevel,
+        )
+
+        self.__transport: transports.AsyncStreamReadTransport = transport
+        self.__recv_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently receving data on this endpoint")
+
+    def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
+        try:
+            transport = self.__transport
+        except AttributeError:
+            return
+        if not transport.is_closing():
+            msg = f"unclosed endpoint {self!r} pointing to {transport!r} (and cannot be closed synchronously)"
+            _warn(msg, ResourceWarning, source=self)
+
+    def is_closing(self) -> bool:
+        """
+        Checks if the endpoint is closed or in the process of being closed.
+
+        Returns:
+            :data:`True` if the endpoint is closed.
+        """
+        return self.__transport.is_closing()
+
+    async def aclose(self) -> None:
+        """
+        Closes the endpoint.
+        """
+        await self.__transport.aclose()
+        self.__receiver.clear()
+
+    async def recv_packet(self) -> _T_ReceivedPacket:
+        """
+        Waits for a new packet to arrive from the remote endpoint.
+
+        Raises:
+            ConnectionAbortedError: the read end of the stream is closed.
+            StreamProtocolParseError: invalid data received.
+
+        Returns:
+            the received packet.
+        """
+        with self.__recv_guard:
+            receiver = self.__receiver
+
+            return await receiver.receive()
+
+    @_utils.inherit_doc(transports.AsyncBaseTransport)
+    def backend(self) -> AsyncBackend:
+        return self.__transport.backend()
+
+    @property
+    def max_recv_size(self) -> int:
+        """Read buffer size. Read-only attribute."""
+        return self.__receiver.max_recv_size
+
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
+
+
+class AsyncStreamSenderEndpoint(transports.AsyncBaseTransport, Generic[_T_SentPacket]):
+    """
+    A write-only communication endpoint based on continuous stream data transport.
+    """
+
+    __slots__ = (
+        "__transport",
+        "__sender",
+        "__send_guard",
+    )
+
+    def __init__(
+        self,
+        transport: transports.AsyncStreamWriteTransport,
+        protocol: protocol_module.StreamProtocol[_T_SentPacket, Any],
+    ) -> None:
+        """
+        Parameters:
+            transport: The data transport to use.
+            protocol: The :term:`protocol object` to use.
+        """
+
+        if not isinstance(transport, transports.AsyncStreamWriteTransport):
+            raise TypeError(f"Expected an AsyncStreamWriteTransport object, got {transport!r}")
+
+        self.__sender: _DataSenderImpl[_T_SentPacket] = _DataSenderImpl(transport, _stream.StreamDataProducer(protocol))
+
+        self.__transport: transports.AsyncStreamWriteTransport = transport
+        self.__send_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently sending data on this endpoint")
+
+    def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
+        try:
+            transport = self.__transport
+        except AttributeError:
+            return
+        if not transport.is_closing():
+            msg = f"unclosed endpoint {self!r} pointing to {transport!r} (and cannot be closed synchronously)"
+            _warn(msg, ResourceWarning, source=self)
+
+    def is_closing(self) -> bool:
+        """
+        Checks if the endpoint is closed or in the process of being closed.
+
+        Returns:
+            :data:`True` if the endpoint is closed.
+        """
+        return self.__transport.is_closing()
+
+    async def aclose(self) -> None:
+        """
+        Closes the endpoint.
+        """
+        await self.__transport.aclose()
+
+    async def send_packet(self, packet: _T_SentPacket) -> None:
+        """
+        Sends `packet` to the remote endpoint.
+
+        Warning:
+            In the case of a cancellation, it is impossible to know if all the packet data has been sent.
+            This would leave the connection in an inconsistent state.
+
+        Parameters:
+            packet: the Python object to send.
+        """
+        with self.__send_guard:
+            sender = self.__sender
+
+            return await sender.send(packet)
+
+    @_utils.inherit_doc(transports.AsyncBaseTransport)
+    def backend(self) -> AsyncBackend:
+        return self.__transport.backend()
+
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
+
+
+class AsyncStreamEndpoint(transports.AsyncBaseTransport, Generic[_T_SentPacket, _T_ReceivedPacket]):
+    """
+    A full-duplex communication endpoint based on continuous stream data transport.
     """
 
     __slots__ = (
@@ -45,12 +239,11 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
         "__send_guard",
         "__recv_guard",
         "__eof_sent",
-        "__weakref__",
     )
 
     def __init__(
         self,
-        transport: transports.AsyncStreamTransport | transports.AsyncStreamReadTransport | transports.AsyncStreamWriteTransport,
+        transport: transports.AsyncStreamTransport,
         protocol: protocol_module.StreamProtocol[_T_SentPacket, _T_ReceivedPacket],
         max_recv_size: int,
         *,
@@ -76,45 +269,22 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
                                                          :exc:`.ManualBufferAllocationWarning`.
         """
 
-        if not isinstance(transport, (transports.AsyncStreamReadTransport, transports.AsyncStreamWriteTransport)):
+        if not isinstance(transport, transports.AsyncStreamTransport):
             raise TypeError(f"Expected an AsyncStreamTransport object, got {transport!r}")
-        if not isinstance(max_recv_size, int) or max_recv_size <= 0:
-            raise ValueError("'max_recv_size' must be a strictly positive integer")
-        if manual_buffer_allocation not in ("try", "no", "force"):
-            raise ValueError('"manual_buffer_allocation" must be "try", "no" or "force"')
+        _check_max_recv_size_value(max_recv_size)
+        _check_manual_buffer_allocation_value(manual_buffer_allocation)
         manual_buffer_allocation_warning_stacklevel = max(manual_buffer_allocation_warning_stacklevel, 2)
 
-        self.__sender: _DataSenderImpl[_T_SentPacket] | None = None
-        if isinstance(transport, transports.AsyncStreamWriteTransport):
-            self.__sender = _DataSenderImpl(transport, _stream.StreamDataProducer(protocol))
+        self.__sender: _DataSenderImpl[_T_SentPacket] = _DataSenderImpl(transport, _stream.StreamDataProducer(protocol))
+        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] | _BufferedReceiverImpl[_T_ReceivedPacket] = _get_receiver(
+            transport=transport,
+            protocol=protocol,
+            max_recv_size=max_recv_size,
+            manual_buffer_allocation=manual_buffer_allocation,
+            manual_buffer_allocation_warning_stacklevel=manual_buffer_allocation_warning_stacklevel,
+        )
 
-        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] | _BufferedReceiverImpl[_T_ReceivedPacket] | None = None
-        if isinstance(transport, transports.AsyncStreamReadTransport):
-            match manual_buffer_allocation:
-                case "try" | "force":
-                    try:
-                        buffered_consumer = _stream.BufferedStreamDataConsumer(protocol, max_recv_size)
-                        if not isinstance(transport, transports.AsyncBufferedStreamReadTransport):
-                            msg = f"The transport implementation {transport!r} does not implement AsyncBufferedStreamReadTransport interface"
-                            if manual_buffer_allocation == "try":
-                                warnings.warn(
-                                    f'{msg}. Consider explicitly setting the "manual_buffer_allocation" strategy to "no".',
-                                    category=ManualBufferAllocationWarning,
-                                    stacklevel=manual_buffer_allocation_warning_stacklevel,
-                                )
-                            raise UnsupportedOperation(msg)
-                        self.__receiver = _BufferedReceiverImpl(transport, buffered_consumer)
-                    except UnsupportedOperation as exc:
-                        if manual_buffer_allocation == "force":
-                            exc.add_note('Consider setting the "manual_buffer_allocation" strategy to "no"')
-                            raise
-                        self.__receiver = _DataReceiverImpl(transport, _stream.StreamDataConsumer(protocol), max_recv_size)
-                case "no":
-                    self.__receiver = _DataReceiverImpl(transport, _stream.StreamDataConsumer(protocol), max_recv_size)
-                case _:  # pragma: no cover
-                    assert_never(manual_buffer_allocation)
-
-        self.__transport: transports.AsyncStreamReadTransport | transports.AsyncStreamWriteTransport = transport
+        self.__transport: transports.AsyncStreamTransport = transport
         self.__send_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently sending data on this endpoint")
         self.__recv_guard: _utils.ResourceGuard = _utils.ResourceGuard("another task is currently receving data on this endpoint")
         self.__eof_sent: bool = False
@@ -142,10 +312,7 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
         Closes the endpoint.
         """
         await self.__transport.aclose()
-        if self.__receiver is not None:
-            self.__receiver.clear()
-        if self.__sender is not None:
-            self.__sender.clear()
+        self.__receiver.clear()
 
     async def send_packet(self, packet: _T_SentPacket) -> None:
         """
@@ -167,9 +334,6 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
 
             sender = self.__sender
 
-            if sender is None:
-                raise UnsupportedOperation("transport does not support sending data")
-
             return await sender.send(packet)
 
     async def send_eof(self) -> None:
@@ -184,14 +348,10 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
             if self.__eof_sent:
                 return
 
-            sender = self.__sender
+            transport = self.__transport
 
-            if sender is None or not isinstance(sender.transport, transports.AsyncStreamTransport):
-                raise UnsupportedOperation("transport does not support sending EOF")
-
-            await sender.transport.send_eof()
+            await transport.send_eof()
             self.__eof_sent = True
-            sender.clear()
 
     async def recv_packet(self) -> _T_ReceivedPacket:
         """
@@ -207,9 +367,6 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
         with self.__recv_guard:
             receiver = self.__receiver
 
-            if receiver is None:
-                raise UnsupportedOperation("transport does not support receiving data")
-
             return await receiver.receive()
 
     @_utils.inherit_doc(transports.AsyncBaseTransport)
@@ -219,10 +376,7 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
     @property
     def max_recv_size(self) -> int:
         """Read buffer size. Read-only attribute."""
-        receiver = self.__receiver
-        if receiver is None:
-            return 0
-        return receiver.max_recv_size
+        return self.__receiver.max_recv_size
 
     @property
     def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
@@ -233,9 +387,6 @@ class AsyncStreamEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPack
 class _DataSenderImpl(Generic[_T_SentPacket]):
     transport: transports.AsyncStreamWriteTransport
     producer: _stream.StreamDataProducer[_T_SentPacket]
-
-    def clear(self) -> None:
-        pass
 
     async def send(self, packet: _T_SentPacket) -> None:
         return await self.transport.send_all_from_iterable(self.producer.generate(packet))
@@ -311,3 +462,49 @@ class _BufferedReceiverImpl(Generic[_T_ReceivedPacket]):
                 pass
 
         raise _utils.error_from_errno(_errno.ECONNABORTED, "{strerror} (end-of-stream)")
+
+
+def _get_receiver(
+    transport: transports.AsyncStreamReadTransport,
+    protocol: protocol_module.StreamProtocol[Any, _T_ReceivedPacket],
+    *,
+    max_recv_size: int,
+    manual_buffer_allocation: Literal["try", "no", "force"],
+    manual_buffer_allocation_warning_stacklevel: int,
+) -> _DataReceiverImpl[_T_ReceivedPacket] | _BufferedReceiverImpl[_T_ReceivedPacket]:
+    # Always exclude this function frame.
+    manual_buffer_allocation_warning_stacklevel += 1
+
+    match manual_buffer_allocation:
+        case "try" | "force":
+            try:
+                buffered_consumer = _stream.BufferedStreamDataConsumer(protocol, max_recv_size)
+                if not isinstance(transport, transports.AsyncBufferedStreamReadTransport):
+                    msg = f"The transport implementation {transport!r} does not implement AsyncBufferedStreamReadTransport interface"
+                    if manual_buffer_allocation == "try":
+                        warnings.warn(
+                            f'{msg}. Consider explicitly setting the "manual_buffer_allocation" strategy to "no".',
+                            category=ManualBufferAllocationWarning,
+                            stacklevel=manual_buffer_allocation_warning_stacklevel,
+                        )
+                    raise UnsupportedOperation(msg)
+                return _BufferedReceiverImpl(transport, buffered_consumer)
+            except UnsupportedOperation as exc:
+                if manual_buffer_allocation == "force":
+                    exc.add_note('Consider setting the "manual_buffer_allocation" strategy to "no"')
+                    raise
+                return _DataReceiverImpl(transport, _stream.StreamDataConsumer(protocol), max_recv_size)
+        case "no":
+            return _DataReceiverImpl(transport, _stream.StreamDataConsumer(protocol), max_recv_size)
+        case _:  # pragma: no cover
+            assert_never(manual_buffer_allocation)
+
+
+def _check_max_recv_size_value(max_recv_size: int) -> None:
+    if not isinstance(max_recv_size, int) or max_recv_size <= 0:
+        raise ValueError("'max_recv_size' must be a strictly positive integer")
+
+
+def _check_manual_buffer_allocation_value(manual_buffer_allocation: Literal["try", "no", "force"]) -> None:
+    if manual_buffer_allocation not in ("try", "no", "force"):
+        raise ValueError('"manual_buffer_allocation" must be "try", "no" or "force"')
