+++ src/easynetwork/lowlevel/api_sync/endpoints/datagram.py
@@ -16,37 +16,39 @@
 
 from __future__ import annotations
 
-__all__ = ["DatagramEndpoint"]
+__all__ = [
+    "DatagramEndpoint",
+    "DatagramReceiverEndpoint",
+    "DatagramSenderEndpoint",
+]
 
+import dataclasses
 import math
 import warnings
 from collections.abc import Callable, Mapping
-from typing import Any, Generic, TypeGuard
+from typing import Any, Generic
 
 from .... import protocol as protocol_module
 from ...._typevars import _T_ReceivedPacket, _T_SentPacket
-from ....exceptions import DatagramProtocolParseError, UnsupportedOperation
-from ... import _utils, typed_attr
+from ....exceptions import DatagramProtocolParseError
+from ... import _utils
 from ..transports import abc as transports
 
 
-class DatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPacket, _T_ReceivedPacket]):
+class DatagramReceiverEndpoint(transports.BaseTransport, Generic[_T_ReceivedPacket]):
     """
-    A communication endpoint based on unreliable packets of data.
+    A read-only communication endpoint based on unreliable packets of data.
     """
 
     __slots__ = (
         "__transport",
-        "__is_read_transport",
-        "__is_write_transport",
-        "__protocol",
-        "__weakref__",
+        "__receiver",
     )
 
     def __init__(
         self,
-        transport: transports.DatagramTransport | transports.DatagramReadTransport | transports.DatagramWriteTransport,
-        protocol: protocol_module.DatagramProtocol[_T_SentPacket, _T_ReceivedPacket],
+        transport: transports.DatagramReadTransport,
+        protocol: protocol_module.DatagramProtocol[Any, _T_ReceivedPacket],
     ) -> None:
         """
         Parameters:
@@ -54,15 +56,96 @@ class DatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPacket,
             protocol: The :term:`protocol object` to use.
         """
 
-        if not isinstance(transport, (transports.DatagramReadTransport, transports.DatagramWriteTransport)):
-            raise TypeError(f"Expected a DatagramTransport object, got {transport!r}")
+        if not isinstance(transport, transports.DatagramReadTransport):
+            raise TypeError(f"Expected a DatagramReadTransport object, got {transport!r}")
         if not isinstance(protocol, protocol_module.DatagramProtocol):
             raise TypeError(f"Expected a DatagramProtocol object, got {protocol!r}")
 
-        self.__is_read_transport: bool = isinstance(transport, transports.DatagramReadTransport)
-        self.__is_write_transport: bool = isinstance(transport, transports.DatagramWriteTransport)
-        self.__transport: transports.DatagramReadTransport | transports.DatagramWriteTransport = transport
-        self.__protocol: protocol_module.DatagramProtocol[_T_SentPacket, _T_ReceivedPacket] = protocol
+        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] = _DataReceiverImpl(transport, protocol)
+
+        self.__transport: transports.DatagramReadTransport = transport
+
+    def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
+        try:
+            transport = self.__transport
+        except AttributeError:
+            return
+        if not transport.is_closed():
+            _warn(f"unclosed endpoint {self!r}", ResourceWarning, source=self)
+            transport.close()
+
+    def is_closed(self) -> bool:
+        """
+        Checks if :meth:`close` has been called.
+
+        Returns:
+            :data:`True` if the endpoint is closed.
+        """
+        return self.__transport.is_closed()
+
+    def close(self) -> None:
+        """
+        Closes the endpoint.
+        """
+        self.__transport.close()
+
+    def recv_packet(self, *, timeout: float | None = None) -> _T_ReceivedPacket:
+        """
+        Waits for a new packet from the remote endpoint.
+
+        If `timeout` is not :data:`None`, the entire receive operation will take at most `timeout` seconds.
+
+        Parameters:
+            timeout: the allowed time (in seconds) for blocking operations.
+
+        Raises:
+            TimeoutError: the receive operation does not end up after `timeout` seconds.
+            DatagramProtocolParseError: invalid data received.
+
+        Returns:
+            the received packet.
+        """
+        receiver = self.__receiver
+
+        if timeout is None:
+            timeout = math.inf
+
+        return receiver.receive(timeout)
+
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
+
+
+class DatagramSenderEndpoint(transports.BaseTransport, Generic[_T_SentPacket]):
+    """
+    A write-only communication endpoint based on unreliable packets of data.
+    """
+
+    __slots__ = (
+        "__transport",
+        "__sender",
+    )
+
+    def __init__(
+        self,
+        transport: transports.DatagramWriteTransport,
+        protocol: protocol_module.DatagramProtocol[_T_SentPacket, Any],
+    ) -> None:
+        """
+        Parameters:
+            transport: The data transport to use.
+            protocol: The :term:`protocol object` to use.
+        """
+
+        if not isinstance(transport, transports.DatagramWriteTransport):
+            raise TypeError(f"Expected a DatagramWriteTransport object, got {transport!r}")
+        if not isinstance(protocol, protocol_module.DatagramProtocol):
+            raise TypeError(f"Expected a DatagramProtocol object, got {protocol!r}")
+
+        self.__sender: _DataSenderImpl[_T_SentPacket] = _DataSenderImpl(transport, protocol)
+
+        self.__transport: transports.DatagramWriteTransport = transport
 
     def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
         try:
@@ -106,23 +189,98 @@ class DatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPacket,
         Raises:
             TimeoutError: the send operation does not end up after `timeout` seconds.
         """
+        sender = self.__sender
+
         if timeout is None:
             timeout = math.inf
 
-        transport = self.__transport
-        protocol = self.__protocol
+        return sender.send(packet, timeout)
 
-        if not self.__supports_write(transport):
-            raise UnsupportedOperation("transport does not support sending data")
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
 
+
+class DatagramEndpoint(transports.BaseTransport, Generic[_T_SentPacket, _T_ReceivedPacket]):
+    """
+    A full-duplex communication endpoint based on unreliable packets of data.
+    """
+
+    __slots__ = (
+        "__transport",
+        "__sender",
+        "__receiver",
+    )
+
+    def __init__(
+        self,
+        transport: transports.DatagramTransport,
+        protocol: protocol_module.DatagramProtocol[_T_SentPacket, _T_ReceivedPacket],
+    ) -> None:
+        """
+        Parameters:
+            transport: The data transport to use.
+            protocol: The :term:`protocol object` to use.
+        """
+
+        if not isinstance(transport, transports.DatagramTransport):
+            raise TypeError(f"Expected a DatagramTransport object, got {transport!r}")
+        if not isinstance(protocol, protocol_module.DatagramProtocol):
+            raise TypeError(f"Expected a DatagramProtocol object, got {protocol!r}")
+
+        self.__sender: _DataSenderImpl[_T_SentPacket] = _DataSenderImpl(transport, protocol)
+        self.__receiver: _DataReceiverImpl[_T_ReceivedPacket] = _DataReceiverImpl(transport, protocol)
+
+        self.__transport: transports.DatagramTransport = transport
+
+    def __del__(self, *, _warn: _utils.WarnCallback = warnings.warn) -> None:
         try:
-            datagram: bytes = protocol.make_datagram(packet)
-        except Exception as exc:
-            raise RuntimeError("protocol.make_datagram() crashed") from exc
-        finally:
-            del packet
+            transport = self.__transport
+        except AttributeError:
+            return
+        if not transport.is_closed():
+            _warn(f"unclosed endpoint {self!r}", ResourceWarning, source=self)
+            transport.close()
 
-        transport.send(datagram, timeout)
+    def is_closed(self) -> bool:
+        """
+        Checks if :meth:`close` has been called.
+
+        Returns:
+            :data:`True` if the endpoint is closed.
+        """
+        return self.__transport.is_closed()
+
+    def close(self) -> None:
+        """
+        Closes the endpoint.
+        """
+        self.__transport.close()
+
+    def send_packet(self, packet: _T_SentPacket, *, timeout: float | None = None) -> None:
+        """
+        Sends `packet` to the remote endpoint.
+
+        If `timeout` is not :data:`None`, the entire send operation will take at most `timeout` seconds.
+
+        Warning:
+            A timeout on a send operation is unusual.
+
+            In the case of a timeout, it is impossible to know if all the packet data has been sent.
+
+        Parameters:
+            packet: the Python object to send.
+            timeout: the allowed time (in seconds) for blocking operations.
+
+        Raises:
+            TimeoutError: the send operation does not end up after `timeout` seconds.
+        """
+        sender = self.__sender
+
+        if timeout is None:
+            timeout = math.inf
+
+        return sender.send(packet, timeout)
 
     def recv_packet(self, *, timeout: float | None = None) -> _T_ReceivedPacket:
         """
@@ -140,31 +298,46 @@ class DatagramEndpoint(typed_attr.TypedAttributeProvider, Generic[_T_SentPacket,
         Returns:
             the received packet.
         """
+        receiver = self.__receiver
+
         if timeout is None:
             timeout = math.inf
 
-        transport = self.__transport
-        protocol = self.__protocol
+        return receiver.receive(timeout)
 
-        if not self.__supports_read(transport):
-            raise UnsupportedOperation("transport does not support receiving data")
+    @property
+    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
+        return self.__transport.extra_attributes
 
-        datagram = transport.recv(timeout)
+
+@dataclasses.dataclass(slots=True)
+class _DataSenderImpl(Generic[_T_SentPacket]):
+    transport: transports.DatagramWriteTransport
+    protocol: protocol_module.DatagramProtocol[_T_SentPacket, Any]
+
+    def send(self, packet: _T_SentPacket, timeout: float) -> None:
         try:
-            return protocol.build_packet_from_datagram(datagram)
+            datagram: bytes = self.protocol.make_datagram(packet)
+        except Exception as exc:
+            raise RuntimeError("protocol.make_datagram() crashed") from exc
+        finally:
+            del packet
+
+        self.transport.send(datagram, timeout)
+
+
+@dataclasses.dataclass(slots=True)
+class _DataReceiverImpl(Generic[_T_ReceivedPacket]):
+    transport: transports.DatagramReadTransport
+    protocol: protocol_module.DatagramProtocol[Any, _T_ReceivedPacket]
+
+    def receive(self, timeout: float) -> _T_ReceivedPacket:
+        datagram = self.transport.recv(timeout)
+        try:
+            return self.protocol.build_packet_from_datagram(datagram)
         except DatagramProtocolParseError:
             raise
         except Exception as exc:
             raise RuntimeError("protocol.build_packet_from_datagram() crashed") from exc
         finally:
             del datagram
-
-    def __supports_read(self, transport: transports.BaseTransport) -> TypeGuard[transports.DatagramReadTransport]:
-        return self.__is_read_transport
-
-    def __supports_write(self, transport: transports.BaseTransport) -> TypeGuard[transports.DatagramWriteTransport]:
-        return self.__is_write_transport
-
-    @property
-    def extra_attributes(self) -> Mapping[Any, Callable[[], Any]]:
-        return self.__transport.extra_attributes
