+++ tests/unit_test/test_async/test_lowlevel_api/test_endpoints/test_stream/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+import pytest
+
+pytest.register_assert_rewrite(f"{__package__}.base")
